requirements.txt
requests==2.25.0
PyYAML==5.3.1
sentry-sdk==0.19.4
lxml==4.6.2
